Refactor 'signerApi' module to 'signer' and streamline error handling, change 'SimpleSigner' to 'Signer'
