+++ src/modules/signer/errors/signerError.ts
@@ -1,0 +1,8 @@
+class SignerError extends Error {
+	constructor(message: string) {
+		super('Signer error: ' + message);
+		this.name = 'SignerError';
+	}
+}
+
+export default SignerError;
