+++ src/modules/signer/functions/signTransaction.ts
@@ -1,6 +1,6 @@
-import SimpleSignerError from '../errors/simple-signer-error';
-import getUrl from '../helpers/get-url';
-import messageHandler from '../helpers/message-handler';
+import SignerError from '../errors/signerError';
+import getUrl from '../helpers/getUrl';
+import messageHandler from '../helpers/messageHandler';
 import { IOperationGroup, ISignMessage } from '../types';
 
 import { NETWORK } from '@/common/types/soroban.enum';
@@ -30,7 +30,7 @@ export default async function signTransaction(
 		return signedXDR;
 	} catch (err: unknown) {
 		if (typeof err === 'string') {
-			throw new SimpleSignerError(err);
+			throw new SignerError(err);
 		}
 
 		throw err;
