+++ src/modules/signer/helpers/openPopup.ts
