--- src/modules/signerApi/errors/simple-signer-error.ts
@@ -1,8 +1,0 @@
-class SimpleSignerError extends Error {
-	constructor(message: string) {
-		super('SimpleSigner error: ' + message);
-		this.name = 'SimpleSignerError';
-	}
-}
-
-export default SimpleSignerError;
